Use goimports if available for make fmt
The VS Code go plugin uses goreturns as the default formatting tool,
which executes goimports on files when they are saved. This can cause
spurious changes to any file touched by the editor, since go fmt does
not change import order.

Change the make fmt rule to check if goimports is installed and use that
instead; if it is unavailable a warning is printed and go fmt is called.

Note: goimports will also execute gofmt, so this change is strictly for
stronger formatting.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>