Fix WATCH_NAMESPACE and manager initialization in webhook deployment
Change manager initialization in webhooks deployment to mirror what's
done in general for controllers. This is necessary since the cache used
by controller-runtime needs to be set to watch all namespaces in order
to be able to find pods in namespaces outside the controller deployment.

Previously, webhooks would fail to get resources outside of the deployed
namespace, despite having the permissions to do so.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>