Fix issue where openshift-oauth routing causes problem on k8s
When a workspace is created with openshift-oauth routing in a kubernetes
cluster, the controller will try to reconcile OAuthClients, which do not
exist as a resource on k8s. This causes a reconcile loop when attempting
to remove a WorkspaceRouting, as the finalizer cannot be removed since
deleting OAuthClients will return an error.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>
(cherry picked from commit ddf4533fb1bb93f4a1455ad34a772e1f1a6f2383)