Add feature to set CHE_MACHINE_NAME env var in plugin containers
Che-machine-exec depends on the CHE_MACHINE_NAME env var to find the
correct container to open a terminal into. This commit adds basic
support, setting the env var to match the container name. It is
potentially a fragile fix, as it's likely that aliases are not handled
correctly.

Feature is gated behind experimental features config option.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>