Remove workaround for kubectl hang when updating resources
kubectl versions with patch versions lower than v1.18.8, v1.17.11
or v1.16.14 will hang when applying our CRDs as a patch (i.e. when
the crds already exist). This is due to an issue in kubectl itself,
see issue

https://github.com/kubernetes/kubernetes/issues/91615

Updating to the latest version of kubectl resolves the hang issue, so
the workaround in the Makefile can be removed

Signed-off-by: Angel Misevski <amisevsk@redhat.com>