Make sure defaulted fields are set in the default dockerimage component
The devfile/api ContainerComponent spec sets default values, e.g. for
sourceMapping. Once a resource is created on the cluster, these defaults
will be applied, and so defaulted fields must have a defined value in
the static spec we use to create resources -- otherwise the controller
should detect a difference and attempt to reconcile it away.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>