Add support for workspacerouting managed by external controller
Instead of failing a workspace when the routingClass is unsupported,
apply an annotation (`controller.devfile.io/external_routing`) to
signify that we expect something else to reconcile the object.

This is to allow an implementation-specific controller to reconcile a
workspace routing and have it still be used by the devworkspace
controller.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>