Use same security context for cleanup pods as we do for deployments
To avoid issues around correct functioning on less secure systems,
explicitly specify the security context used for the cleanup job to
match the one used for the DevWorkspace deployment. Otherwise, systems
like minikube will typically run the cleanup job with elevated
permissions, which won't work on e.g. production OpenShift (where UID is
semi-random and not root)

(squashed)
Fix security context used in k8s to specify root GID

The standard for Kubernetes storage is that writeable resources are
created so that the group write bit is set in file permissions and the
file belongs to the root group. To guarantee workspace containers can
access resources as expected, the GID used for the container must be 0

(squashed)
Fix reconcile looping on OpenShift

Pod SecurityContext is defaulted to a non-nil empty struct instead of
nil, causing a loop on OpenShift

Signed-off-by: Angel Misevski <amisevsk@redhat.com>