Avoid issue with non-writeable subdirectory when cleaning up PVC
Kubernetes creates subpaths in PVCs such that only the leaf node has
g+rwx permissions, i.e. for the volumeMount subpath

    dir1/dir2/dir3

the directories dir1 and dir2 will have 755 permissions, whereas dir3
will have 777 permissions. As a result, non-root UID users will not be
able to remove the dir1/dir2 directories. This is a problem with
common-pvc workspace subpaths, as we typically mount e.g.
<workspace-id>/projects, and so we cannot remove the 'projects'
directory from the PVC once a workspace is deleted.

To work around this issue, we make sure that the first volumeMount
encountered in a pod definition uses the `<workspace-id>` subpath. With
this, k8s will first create the `<workspace-id>` directory (with 777
permissions), and then create subpaths as required. This approach
resolves the *main* issue, but will still fail if a volumeMount uses a
multiple-directory subpath (e.g. `<workspace-id>/my-dir/plugins`).

NOTE: it is crucial that the plain subpath volumeMount is the very first
one encountered, in list order. Otherwise, the `<workspace-id>` subpath
will *not* be updated to 777 permissions.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>