Improve plugins support in flatten library
* Add functionality for detecting reference cycles in plugins (e.g.
plugin A references plugin B which references plugin A), avoiding the
potential to lock the operator
* Add basic functionality for supporting compatibility between plugins
and editors, based on annotations applied to DevWorkspaceTemplates
representing plugins; some work would be required for similar
functionality for plugins specified by e.g. ID.

This is done by building a tree of plugin information as we flatten
plugin imports; each node in the tree stores the information about the
plugin, a node representing the component that imported it, and a list
of nodes representing plugins it imports. We can walk this tree to check
e.g. plugin labels for all plugins included in the workspace (to check
compatiblity) or to look for import loops (by following parent
references).

Signed-off-by: Angel Misevski <amisevsk@redhat.com>