Ensure we don't update devworkspace with flattened content in reconcile
Since the reconcile loop has to work on the flattened devworkspace
contents while updating annotations on the original object, some care is
required. To this end:

* Make a copy of the original devworkspace from the cluster, which is
the only object to be updated
* Rework startup timing functionality to not involve directly modifying
the devworkspace resource; instead use a map that we merge into the
(cluster) workspace annotations only when needed.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>