Fix issue where volumeMounts were being unnecessarily added
Improve check for whether the volumeMount subpaths workaround is
required for a particular deployment. Previously, we only checked
whether any volumes were defined, which failed on OpenShift due to a
secret volume being in the list. Now we check to make sure the common
PVC volume is present in the deployment before applying the workaround.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>