Enable the solvers to do the finalization of a workspace routing themselves.
For this to make sense, the openshift oauth solver is now in charge of
managing the oauthclient for the workspace. As such the oauthclient is no
longer part of the ResolvedObjects struct.

The motivation for doing this is the need of the devworkspace-che-operator
to do a custom finalization of the routing - it needs to delete
the configmaps for the gateway configuration which need to live in
a different namespace than the workspace.