Set webhooks to always true, since they're required now
Since we moved to multi-version devfile/api CRDs, the webhook server is
a required component of the DevWorkspace Operator. Currently, the server
is set up to not interfere with any default flows (e.g. local debugging)
as it is a separate deployment. It has also been restricted to only
apply when the restricted-access annotation is set.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>