Fix endpoint specifications for web-terminal plugins
Endpoints for the web terminal plugin had not been updated to match an
update to the devfile/api dependency, which moved attributes protocol
and secure to fields on the Endpoint object and introduced the exposure
field.

This commit also updates how the cluster routingSolver handles endpoints
to match the semantics of endpoint.Exposure.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>