Remove need for componentDescriptions except for che-rest-apis sidecar
Restrict the need to generate ComponentDescriptions so that they're only
used in generating the che-rest-apis sidecar. This allows us to remove
the legacy component descriptions struct once we remove che-rest-apis.

To support components that are not defined in the devfile, make the
generate component descriptions library function ignore containers that
don't come from the current DevWorkspace spec.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>