Refactor provisioning storage to an interface
To support multiple types of workspace storage, we need a separate
interface that handles provisioning persistent storage, rewriting
workspace containers/volumes, and cleaning up the PVC.

This commit introduces the "Provisioner" interface, which defines
methods for
1. Checking if storage is needed for the workspace
2. Provisioning storage and rewriting container volumes/volumeMounts
3. Performing any required cleanup for the configured storage type

This removes all involvement in managing storage from the main
controller loop, instead delegating it to this interface.

The type of storage used for a DevWorkspace is defined by the value of
the 'controller.devfile.io/storage-type' label (which should perhaps be
converted to an attribute once global attributes are supported). If the
attribute is undefined, the common storage strategy is used.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>