Limit async workspaces to one and implement basic cleanup
Limit the number of started async workspaces to one, to avoid some
issues when a workspace is deleted while another workspace is using
async storage.

Implement a basic storage finalizing process for async workspaces, that
scales the async server to 0 before running the common PVC cleanup job.
This would fail when another workspace is currently running.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>