Avoid updating webhooks serviceaccount unless necessary.
Each time the webhooks serviceAccount is updated with nil .secrets, a new secret is
generated to fill the `secrets` field. This results in every restart of
the DevWorkspace operator creating a new secret on the cluster.

Since the only fields we set on the spec ServiceAccount are labels, we
only need to update when the labels differ from the cluster.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>