Resolve git commit ID from local file if .git dir isn't present
In mid/downstream repos, the .git directory may by higher in the
directory tree, resulting in commands that resolve the commit SHA-1
failing.

To avoid this, attempt to resolve the git commit hash in order:
1. Check .git/HEAD and .git/refs and use that value if available
2. Otherwise, use value in .commit_hash file in root directory
3. Otherwise, use "unknown"

Also, to make things simpler, break out build information variables into
version.mk and include it in the main Makefile.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>