Introduce phase to represent failed workspaces that need to be stopped
To avoid issues around starting workspaces with `phase: Failed`,
introduce an intermediate phase "Failing" to represent a workspace that
has hit a failure condition and needs to be stopped.

When the controller encounters a DevWorkspace with `phase: Failing` and
`.spec.started: true`, it will stop that workspace and update the phase
to `Failed`.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>