Add deploy templates for OLM generation that include CSV
* Add deploy/templates/components/csv to be able to include a base CSV
when generating deployment files
* Add deploy/templates/olm to support generating deployment files with a
CSV included.

The `deploy/templates/olm` target is similar to `service-ca` and
`cert-manager` in that it should be used with kustomize to generate
everything needed to deploy the operator. This branch needs to be
separate since we don't want to include a CSV in the standard deployment
files, as they are used to deploy the operator without OLM.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>