Restrict which namespaces DWTs can be read from via annotation
Restrict reading DWTs on the cluster to only those in the same
namespace as the DevWorkspace by default. Additional namespaces can be
enabled by applying an annotation to the DWT, with options for a
comma-separated list of approved namespaces, or '*' to allow the DWT to
be read from anywhere on the cluster.

This is required, as otherwise the privilege to create DevWorkspaces is
allows reading DevWorkspaceTemplates anywhere on the cluster.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>