Move conversion webhooks to main controller deployment
Move conversion webhooks from the standalone webhooks server into the
main controller deployment. This helps in using automatic configuration
from OLM in creating webhooks, as this is unavailable for deployments
not in the CSV.

Using OLM to create the webhook configs and inject service
name/namespace into the CRDs enables deploying the DevWorkspace
controller to any namespace instead of restricting it to
openshift-operators.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>