Add support for workspaceEnv via an attribute on components
Add support for workspaceEnv (i.e. environment variables provided by a
component that get added to all other components) though setting an
attribute 'workspaceEnv' in a component. If that component is added to
the workspace (either directly or through a plugin) the environment
variables defined within are added to all containers and init containers
in the workspace.

Example: a component

  components:
  - name: my-component
    attributes:
      workspaceEnv:
        - name: MY_ENV
          value: my-value
        - name: MY_ENV_2
          value: my-value-2

will add the MY_ENV and MY_ENV_2 environment variables to all containers
in the DevWorkspace.

Note we use component attributes rather than top-level attributes for
workspaceEnv, since top-level attributes pose some problems for
flattening:
- If the base DevWorkspace defines workspaceEnv in top-level attributes
  and a plugin does the same, flattening fails
- If two plugins define top-level workspaceEnv attributes, the
  workspaceEnv of the first if overwritten by the second

Signed-off-by: Angel Misevski <amisevsk@redhat.com>