Enable config to log diff on objects when updating
Use the experimental features flag to enable additional logging when
updating objects on the cluster. If experimental features are enabled
and a spec object is found to be different from the cluster one, log a
diff between the two. This can potentially be useful for diagnosing
reconcile loops.

Due to json-structured logging, the formatted message is escaped into a
string, and so must be parsed for reading (e.g. by piping through sed:

    make run 2>&1 | sed -e 's|\\n|\n|g; s|\\t|    |g'

for quick checking, the .msg can be copy-pasted as the argument the
printf to interpret \t and \n

Signed-off-by: Angel Misevski <amisevsk@redhat.com>