Workaround issue where makefile can't run commands after oc logout
Executing `oc logout` drops user info from kubeconfig. This does not
pose a problem for `oc` as it understands what it has done, but kubectl
will prompt for login info when executing commands.

This causes any make rule to hang if the kubectl is installed, as
deploy.mk is imported and contains a check for kubectl api-resources.
Since this call will try to read a password from stdin, the call will
never return and so the comparison will hang silently.

This commit closes stdin on all kubectl/oc commands.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>