Add functionality to merge duplicate volume components while flattening
Merge volume components to avoid issues where e.g. a devworkspaces defines a
volume that is also defined in a plugin. This makes devworkspaces more
portable, as adding a plugin is less likely to break an existing
workspace.

Volumes are merged so that the only instance is found in the
devworkspace, its parent, or its plugins in that order (i.e. volumes in
plugins will be deleted in favor of a volume defined in the devworkspace
itself).

To try to avoid issues:
* If one duplicated volume is not ephemeral, the merged volume is not
  ephemeral.
* If two volumes define different sizes, the merged volume uses the
  larger size.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>