Propagate DevWorkspace .spec.started status to an annotation on routings
Add annotation 'controller.devfile.io/devworkspace-started' to be
applied to DevWorkspaceRoutings. If a workspace is stopped, the
annotation is set to "false", otherwise it is "true". If the
DevWorkspaceRouting reconciler encounters a DevWorkspaceRouting with the
annotation set to false, it ends the reconcile early.

This ensures that every workspace start/stop event triggers a
reconcile, and allows the routing to clean up resources if necessary.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>