Restrict secrets and configmaps in the cache via new labels
Add labels

  - controller.devfile.io/watch-configmap
  - controller.devfile.io/watch-secret

which must be set to "true" in order for the DevWorkspace Operator to
see the  corresponding secret/configmap.

This is required (compare to the previous commit) because the controller
is not only interested in secrets and configmaps it creates, but also
any configmap/secret on the cluster with e.g. the automount label
attached.

Since each type in the controller gets a single informer, we can only
specify a single label selector for the objects we are interested in.
This means we cannot have e.g. "has devworkspace_id label OR has
mount-to-devworkspace label".

Signed-off-by: Angel Misevski <amisevsk@redhat.com>