Update how workspace metadata configmap is handled
Restricting the cache to only configmaps with the new label results in
existing workspaces failing to reconcile. This occurs because attempting
to Get() the configmap from the cluster returns a IsNotFound error,
whereas attempting to Create() the configmap returns an AlreadyExists
error (Create interacts with the cluster, Get interacts with the cache).

To avoid this, if we encounter an AlreadyExists error when attempting to
create an object, we optimistically try to update the object (thus
adding the required label). This resolves the issue above, as if the
obejct is updated, the subsequent Get() call will return the object as
expected.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>