Collect sync code into one interface to simplify syncing with cluster
Restricting the controller-runtime cache to specific objects means that
once-tracked objects can disappear from the controller's knowledge if
the required label is removed. To work around this, it is necessary to
update how we sync objects to specifically handle the case where:

  * client.Get(object) returns IsNotFound
  * client.Create(object) returns AlreadyExists

This occurs because we can't read objects that aren't in the cache, but
attempting to create objects collides with the actual object on the
cluster.

Since the basic flow of Get -> Create/Update is repeated for each type
we handle, this commit collects that repeated logic into one package
(pkg/provision/sync), allowing object handling to be done in one place.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>