Treat updating services specially since we have to copy ClusterIP
For most objects, we can client.Update() using the spec object without
issue. However, for Services, updates are rejected if they try to unset
spec.ClusterIP. This means we need to copy the ClusterIP from the
cluster service before updating.

This commit adds an extensible mechanism for specifying type-specific
update functions that are called whenever we attempt to update a cluster
object.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>