Treat all 4xx status codes as "unimplemented" when doing server check
Some editors return HTTP 400 for the /healthz endpoint, breaking the
server readiness check and leaving DevWorkspaces in a "Waiting for
editor to start" state forever. Since health checks are kind of flaky to
begin with, this commit relaxes the check to treat any 4xx response code
as "unimplemented or behind auth" in order to not block workspace
startup.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>