Add check that .spec.replaces is added to the CSV before release
Downstream catalogs have not yet migrated to the file-based catalog
system in newer versions of OLM, so the .spec.replaces field is
mandatory in releases. This field must be set manually for now, as it's
difficult to determine the correct version to use in an automated way.

This commit causes `make-release.sh --release` to fail if this field is
not present in the current branch.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>