Revert to using opm v1.18.z way of building indexes
Building file-based catalogs is broken in opm v1.19.1 as `opm render`
does not produce valid output. In order for built indexes to properly
show operators in the OperatorHub UI, we're forced to use the previous
method (`opm index add ...`) instead. This commit automates adding
replaces fields to the CSVs in order to allow using the previous
functionality.

On release, new bundles are added to the index optimistically; this
could fail if e.g. the resulting upgrade graph isn't clean and means
that running the release script twice for a particular release will
fail.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>