Make project-clone work in a temporary directory; copy files at end
To avoid issues where the project-clone container is killed while
cloning (and thus leaving repos in an inconsistent state), do all
actions to a temporary directory and move files into their expected
location only as a last step. This should allow projects to be cloned
correctly if the project-clone container is killed early in execution.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>