Use endpoint name as container port name if it is short enough
Reuse the endpoint name as the port name in a workspace container if it
fits the pod spec for a port name (max length = 15 chars). If an endpoint
name is longer than 15 chars, fallback to previous functionality of
formatting <portNum>-<protocol>. In the case of duplicated ports in one
container with short names, use the name from the first port.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>