Use git binary for checking out references to avoid go-git issues
Instead of using the go-git library for checking out tags, commits, and
branches, delegate these tasks to the git binary installed in project
clone.

In some cases, go-git will fail to checkout references if a repository
is large and/or if the changeset between the main and the reference is
significant.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>