Support removing all workspace-owned objects on stop via config
Add support for removing all objects owned by the DevWorkspace when it
is stopped. Feature is disabled by default and can be enabled by setting
config value in the configuration CR.

When a workspace is stopped, remove all deployments, secrets,
configmaps, and DevWorkspaceRoutings with the label

controller.devfile.io/devworkspace_id=<workspaceID>

Removing the DevWorkspaceRouting ensures that all services, routes, and
ingresses are removed as well.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>