Set fsGroup in pod security context to 1234 instead of leaving empty
In some Kubernetes clusters, fsGroup needs to be set in order for
workspace containers to read contents of mounted PVCs. This change
matches the fsGroup with the runAsUser field. Note this is the security
context that is used in Eclipse Che as well.

Only applies to Kubernetes; behavior in OpenShift is unchanged.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>