Don't add project-clone when there are no containers in DevWorkspace
Avoid a cryptic error in starting a DevWorkspace where the only
container/initContainer is project-clone. Since the project-clone init
container by default assumes that it will be mounting the projects
volume, attaching it to a DevWorkspace with no containers will result in
an invalid deployment since the project-clone initContainer references a
non-existent volume.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>