Ignore FailedMount events if they're due to serving certs
In some cases, the workspace pod can be created before the service-ca
oeprator on OpenShift creates the serving cert secret, resulting in
temporary FailedMount events occurring for workspaces that depend on
that functionality (e.g. routingClass: cluster-tls or web-terminal).

To avoid this, we ignore those events regardless of whether or not
they're in the ignoredUnrecoverableEvents config setting.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>