[project-clone] Avoid edge case where revision is default branch
Fix error case where a project checkoutFrom defines a revision that
matches the default branch for a repo (e.g. "main"). In this case, the
branch already exists and so `git checkout -b` will fail.

Instead, if the branch exists in the local repo, project-clone will
checkout that branch and update the tracking branch to match the remote.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>