Ignore order of envFrom when comparing deployments
When listing configmaps and secrets on the cluster, the API can return
objects in a random order (for two configmaps, I see the reverse order
almost exactly 1/8th of the time)

This can result in the operator defining a Deployment spec for
workspaces with the envFrom fields switched, if those configmaps are
automounted as environment variables. To get around this, we ignore the
order of envFrom entries when deciding if the deployment needs to be
updated.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>