Pause workspace startup if shared PVC is being deleted
In order to avoid a situation where a starting workspace attempts to
mount a terminating PVC, preventing the PVC from being deleted and the
workspace from starting, we pause workspace startup if the PVC to be
used is terminating. Once the PVC is removed from the cluster, it is
recreated and workspace startup may proceed.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>