Fix ignoring order of volumeMounts when comparing deployments
Since support for subpath automount volumes was added, the comparison
function used to ignore the order of volumeMounts in a deployment is no
longer valid, as multiple volumeMounts can use the same volume. This can
result in spurious changes to the deployment's spec, if the operator
reads volumes in a random order.

To fix this, also compare mountPath and subPath. The comparison function
for other slices is unchanged, as in those cases the name is a key in
the list and must be unique.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>