Fix issue where deleting a workspace deletes the PVC incorrectly
Fix a regression caused by commit cbd3fd11. When deleting workspaces, if
there are no non-terminating workspaces left in the namespace, we delete
the common PVC rather than running cleanup jobs. However, the way the
check was written assumed that we were counting _all_ workspaces, and
would delete the common PVC if there were two workspaces in a namespace
and one was deleted.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>