Avoid logging error from updating status of a deleted workspace
When the last finalizer on a DevWorkspace is cleared, the cluster may
garbage-collect that workspace immediately, which can result in the
object being deleted before we reach the point of attempting to update
the workspace's status. This results in logging a harmless but confusing
error, so we don't try to update the status for terminating workspaces
with no finalizers.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>
(cherry picked from commit e3805fb8c19d95c5efa5e9ecf07cabe0c5719f29)