Queue reconcile for all common storage-type workspaces on PVC deletion
To make sure workspaces that are in an errored state due to PVC cleanup
failing are removed when the common PVC is deleted (e.g. when all
DevWorkspaces are deleted), add a watch and enqueue reconciles for all
common-storage workspaces when the common PVC is deleted.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>
(cherry picked from commit a2e149f5b147adf70861ba6aecc704545eca21fa)