Limit objects cached in webhooks manager
When validating pods/exec requests, due to how Kubernetes serves
webhook requests, the webhook server has to read the pod being execed
into from the cluster. Since it does this using the manager, it's
possible for the manager's internal cache to grow too large for the
default webhook server memory limit, causing it to be killed.

If this occurs, the cluster is left in a bad state until either the
number of pods on the cluster is reduced or the webhooks server memory
limit is increased.

This PR adds a cache filter similar to the main controller to limit
which pods the webhook server caches.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>