Block changes to container images in restricted-access pods
For some reason, it's possible to edit a pod controlled by a replicaset
to use a different image, and that change will not be reverted by the
replicaset.

We don't want to allow this for restricted-access pods, as it could
potentially allow someone to silently replace the image used for a
workspace pod.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>