Introduce DevWorkspaceWithConfig type to store scoped workspace config
Introduce new struct that stores both the current DevWorkspace being
worked on and the config when reconcile began. This struct will be used
in place of the plain workspace where possible in order to allow
per-workspace configuration through the DWOC.

This commit breaks the Go build as it does not update any files that
depend on the workspace.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>
Co-authored-by: Andrew Obuchowicz <aobuchow@redhat.com>