feat: Allow external DWOC to be merged with internal DWOC
Allow specifying an external DWOC that gets merged with the global DWOC for use by a workspace.
During the merge, the fields which are set in the external DWOC will overwrite those existing in the internal/global DWOC,
resulting in a merged DWOC to be used by the workspace. The internal/global DWOC will remain unchanged after the merge.

The external DWOC's name and namespace are specified with the `controller.devfile.io/devworkspace-config` DevWorkspace attribute,
which has the following structure:

attributes:
  controller.devfile.io/devworkspace-config:
    name: <string>
    namespace: <string>

Part of eclipse/che#21405

Signed-off-by: Andrew Obuchowicz <aobuchow@redhat.com>