Add git-lfs section when automount configmap only mounts gitconfig
Always create a 'merged' git config in order to insert git-lfs
configuration, even if only a /etc/gitconfig automount is specified.
Previously, merging was only done when we needed to combine multiple
sources (e.g. with git-credentials) or when we need to create a
gitconfig from scratch. This meant that attempting to use a custom
gitconfig would overwrite the git-lfs section unless the gitconfig
contained it.

The reason to add it even when a user explicitly defines a gitconfig is
that generally a user's gitconfig is stored in $HOME/.config while the
git-lfs configuration is stored in /etc/gitconfig. For DevWorkspaces, we
can only use /etc/gitconfig.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>