Use server-side apply for make install
Since updating the DevWorkspace and DevWorkspaceTemplate CRDs, it is no
longer possible to use `kubectl apply` to create the object in a
cluster, due to the 'last-applied-configuration' annotation becoming too
long for an annotation.

To avoid this, we pass --server-side=true in kubectl apply commands.
Since the CA bundle injection in CRDs results in a different manager for
the .spec.conversion.webhook.clientConfig.caBundle field, we have to
pass --force-conflicts in `make install` to be able to re-apply the CRDs
if they are already applied.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>