Check that containerContributions attributes can be parsed
When the 'controller.devfile.io/container-contribution' or
'controller.devfile.io/merge-contribution' attribute is present on a
component in a DevWorkspace, check whether its value can be parse as
a boolean. If the value cannot be read as a boolean (e.g. a typo), then
fail workspace start with a message pointing at the issue.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>