Regenerate DeepCopy files for CRDs
Regenerate DeepCopy files for CRDs after switching how controller-gen is
downloaded. Previously, we would download + build controller-gen locally
and use that in scripts. Now we `go install` it to ./bin and use that,
which is a lot simpler.

However, despite both instances using controller-gen v0.6.1, the output
of deepcopy-gen is different:

Previously, locally-built controller gen included the line
//go:build !ignore_autogenerated
at the top of deepcopy files

The new, `go install`ed version omits this line.

¯\_(ツ)_/¯

Signed-off-by: Angel Misevski <amisevsk@redhat.com>