Stop adding workspace SA to SCC and don't use SA finalizer
Since SCCs are added to serviceaccounts via roles and rolebindings, it's
no longer necessary to update SCCs to add workspace serviceaccounts as
users. This step is removed, as is functionality for automatically
adding the SA finalizer to workspaces.

However, code related to finalizing workspaces and cleaning up changes
made to SCCs is kept intact in order to ensure any workspaces that
previously used the finalizer can still be deleted cleanly.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>