Update sync diff options to avoid common loops with k8s components
As DevWorkspaces can define Kubernetes objects in components, it's
necessary to improve the checks we do to determine if an object on the
cluster needs to be updated. This is especially an issue in the case of
defaulted fields in objects, as comparing in a straightforward way will
mean the object has to be updated every single time.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>