Extend kube component webhooks to DevWorkspaceTemplates as well
Add webhook checks to prevent creation/mutation of DevWorkspaceTemplates
that define Kubernetes components if the requesting user does not have
those permissions. This prevents working around RBAC checks in
DevWorkspaces by just creating a DevWorkspaceTemplate instead.

Note however that currently DWO does not check that a user creating a
workspace has permissions to reference this DevWorkspaceTemplate -- if a
user can reference a DWT, they will be able to create the objects
defined in that DWT via a DevWorkspace without further RBAC
verification.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>